disable plotly resize handler to keep better plot sizes
It's not ideal as it would be better to enable it to make plots
responsive, but the current setting basically messes up
the plot sizes whenever I navigate away from the page then
come back.

Testing to see if this works better. Need to keep an eye
on phone portrait/landscape switches as corner case.

Signed-off-by: Gergely Imreh <gergely.imreh@faculty.ai>